Load color names and data from an external file.
... instead of being in a header file. Makes it much easier to add your
own colors or otherwise modify the list.
