+++ utils/genrgb.pl
@@ -1,14 +1,14 @@
 #!/usr/bin/perl
 #
 # Script for generating the color name to RGB component maps in a big
-# json string.
+# json string. Run from the root pennmush directory, not utils.
 #
 # Gets xterm color data from utils/xterm256.json and other color
 # schemes from Graphics::ColorNames
 #
 # Requires a fairly modern perl and the File::Slurper,
 # Graphics::ColorNames and JSON::XS packages. Install off of CPAN or
-# your package manager (lib-file-slurper-perl,
+# your package manager (libfile-slurper-perl,
 # libgraphics-colornames-perl, libjson-xs-perl on Debian/Ubuntu/etc.,
 # perl-file-slurper, perl-graphics-colornames, perl-json-xs on arch)
 #
@@ -131,7 +131,6 @@ foreach my $color (@$xterm256colors) {
     push @allcolors, {name => "xterm$num", rgb => hex $rgb, xterm => int $num, ansi => int $ansi};
 }
 
-
 while (my ($name, $rgb) = each %x11colors) {
     $name =~ s/\s+//;
     $name = lc $name;
@@ -157,64 +156,10 @@ foreach my $color (@$xterm256colors) {
     push @allcolors, {name => $name, rgb => hex $rgb, xterm => int $num, ansi => int $ansi};
 }
 
-open OUTPUT, ">", "hdrs/rgb.h";
-print OUTPUT <<EOT;
-/* rgb.h
- *
- * Generated by $0, this file contains mappings of color names to RGB and ANSI vaules.
- * Should only be used by markup.c.
- */
+@allcolors = sort { $a->{name} cmp $b->{name} } @allcolors;
 
-EOT
-
-say OUTPUT 'const char *colors_json[] = {';
-
-my $printer = JSON::XS->new->ascii->pretty(0)->canonical;
-while (scalar @allcolors) {
-    my @chunk = splice @allcolors, 0, 70;
-    my $output = $printer->encode(\@chunk);
-    $output =~ s/"/\\"/g;
-    say OUTPUT "    \"$output\",";
-}
-
-say OUTPUT "    NULL\n};\n";
-
-say OUTPUT <<EOC;
-/** Info on a color from the 16-color ANSI palette */
-struct COLORMAP_16 {
-  int id; /**< Code for this color (0-7) */
-  char desc; /**< Lowercase char representing color */
-  uint8_t hilite; /**< Is this char highlighted? */
-  uint32_t hex; /**< Hex code for this color */
-};
-
-/* Taken from the xterm color chart on wikipedia */
-struct COLORMAP_16 colormap_16[] = {
-  /* normal colors */
-  {0, 'x', 0, 0x000000},
-  {1, 'r', 0, 0xcd0000},
-  {2, 'g', 0, 0x00cd00},
-  {3, 'y', 0, 0xcdcd00},
-  {4, 'b', 0, 0x0000ee},
-  {5, 'm', 0, 0xcd00cd},
-  {6, 'c', 0, 0x00cdcd},
-  {7, 'w', 0, 0xe5e5e5},
-
-  /* Hilite colors */
-  {0, 'x', 1, 0x7f7f7f},
-  {1, 'r', 1, 0xff0000},
-  {2, 'g', 1, 0x00ff00},
-  {3, 'y', 1, 0xffff00},
-  {4, 'b', 1, 0x5c5cff},
-  {5, 'm', 1, 0xff00ff},
-  {6, 'c', 1, 0x00ffff},
-  {7, 'w', 1, 0xffffff},
-
-  {-1, 0, 0, 0}
-};
-EOC
-
-close OUTPUT;
-
+open my $output, ">", "game/txt/colors.json";
+my $printer = JSON::XS->new->utf8->pretty->canonical;
+print $output $printer->encode(\@allcolors);
+close $output;
 say "Outputted $counter colors.";
-
